Feature/sensor fix

+++ telemetry_device/Statistics/CompactCollection/StatisticsDictionaryValue.cs
@@ -5,8 +5,8 @@ namespace telemetry_device.Statistics.CompactCollection
     class StatisticsDictionaryValue
     {
         public StatisticsSeverity Sevirity { get; set; }
-        public float Value { get; set; }
-        public StatisticsDictionaryValue(StatisticsSeverity sevirity, float value)
+        public double Value { get; set; }
+        public StatisticsDictionaryValue(StatisticsSeverity sevirity, double value)
         {
             Sevirity = sevirity;
             Value = value;
